// src/types/book.ts
export interface Book {
  bookId: number;
  bookName: string;
  author: string;
  image: string;
  review: string;
  totalPages: number;
  rating: number;
  category: string;
  tags: string[];
  publisher: string;
  yearOfPublishing: number;
}

export interface ApiResponse<T> {
  status: string;
  message: string;
  data: T;
}